refactor: active_pillars and active_sentinels features
