+++ lib/rearchitecture/features/active_pillars/active_pillars.dart
@@ -1,3 +1,2 @@
 export 'cubit/active_pillars_cubit.dart';
 export 'view/active_pillars_card.dart';
-export 'widgets/widgets.dart';
