+++ lib/rearchitecture/features/active_pillars/cubit/active_pillars_state.dart
@@ -1,16 +1,16 @@
 part of 'active_pillars_cubit.dart';
 
-/// This state is used by the [ActivePillarsCubit] to track and update the number of
-/// active active_pillars.
+/// This state is used by the [ActivePillarsCubit] to track and update the
+/// number of active active_pillars.
 ///
-/// [ActivePillarsState] stores an integer value representing the number of active_pillars
-/// retrieved from the Zenon network.
+/// [ActivePillarsState] stores an integer value representing the number of
+/// active_pillars retrieved from the Zenon network.
 @JsonSerializable(explicitToJson: true)
 class ActivePillarsState extends TimerState<int> {
   /// Constructs a new [ActivePillarsState] object.
   ///
-  /// The [data] field in this case represents the count of active active_pillars on
-  /// the Zenon network.
+  /// The [data] field in this case represents the count of active
+  /// active_pillars on the Zenon network.
   const ActivePillarsState({
     super.status,
     super.data,
