+++ lib/rearchitecture/features/active_sentinels/active_sentinels.dart
@@ -1,3 +1,2 @@
 export 'cubit/active_sentinels_cubit.dart';
 export 'view/active_sentinels_card.dart';
-export 'widgets/widgets.dart';
